ai-voice/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
openai-whisper==20231117
edge-tts==6.1.7
httpx==0.26.0
python-multipart==0.0.6
